requirements.txt
fastapi==0.111.0
httpx==0.27.0
numpy==1.26.4
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-json-logger==2.0.7
rdkit-pypi
sentence-transformers==2.7.0
sqlalchemy==2.0.30
tiktoken==0.6.0
torch==2.3.0
uvicorn[standard]==0.29.0
tenacity==8.2.3
biopython==1.83
tenacity
pytest-asyncio